A few cleanings

+++ frame/support/src/weights.rs
@@ -50,7 +50,7 @@ use sp_runtime::{
 pub use sp_runtime::transaction_validity::TransactionPriority;
 
 /// Numeric range of a transaction weight.
-pub type Weight = u64;
+pub type Weight = u32;
 
 /// Means of weighing some particular kind of data (`T`).
 pub trait WeighData<T> {
